p50(nvim): Changed Conform's `format_on_save` lambda
allows you to enable a formatter for langages in the disable_filetypes
table to have a formatter that can be run manually with Leader-f but
doesn't enable format_on_save for them. Before this change, if you add
a formatter for a language such as C, like clang-format, it enables
`format_on_save` since the formatter is set explicitly.

See: https://github.com/nvim-lua/kickstart.nvim/pull/1395

Co-authored-by: RulentWave <49258216+RulentWave@users.noreply.github.com>