p50(bpswm): Added more bspwm desktop
Note that in sxhkd, when using key symbol ranges like {1-9,0,-,=}, it
does not automatically translate symbols like - and = to the correct key
names or keycodes when used in brace expansions. This works for
alphanumeric keys but gets tricky with symbols.

So, bindings "-" and "=" separately outside of the brace expansion is
needed.