p50(redis): Replace redis with valkey
Valkey will be replacing redis due to Redis modifying its license
from BSD-3-Clause to RSALv2 and SSPLv1 on March 20th, 2024.

The redis package will be moved to the AUR and will not receive any
additional updates and should be considered deprecated until it is removed.

see: https://archlinux.org/news/valkey-to-replace-redis-in-the-extra-repository/